+++ app/src/main/java/com/koalasat/pokey/MainActivity.kt
@@ -1,13 +1,10 @@
 package com.koalasat.pokey
 
 import android.Manifest
-import android.app.Activity
 import android.content.pm.PackageManager
 import android.os.Build
 import android.os.Bundle
-import android.util.Log
 import android.widget.Toast
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
@@ -18,10 +15,6 @@ import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.koalasat.pokey.databinding.ActivityMainBinding
 import com.koalasat.pokey.models.EncryptedStorage
-import com.vitorpamplona.quartz.signers.ExternalSignerLauncher
-import com.vitorpamplona.quartz.signers.SignerType
-import java.util.UUID
-import kotlin.coroutines.cancellation.CancellationException
 
 class MainActivity : AppCompatActivity() {
     private val requestCodePostNotifications: Int = 1
@@ -62,8 +55,6 @@ class MainActivity : AppCompatActivity() {
                 requestCodePostNotifications,
             )
         }
-
-        if (EncryptedStorage.pubKey.value.isNullOrEmpty()) connectExternalSigner()
     }
 
     override fun onRequestPermissionsResult(
@@ -78,40 +69,6 @@ class MainActivity : AppCompatActivity() {
             } else {
                 Toast.makeText(applicationContext, getString(R.string.permissions_required), Toast.LENGTH_SHORT).show()
             }
-        }
-    }
-
-    private fun connectExternalSigner() {
-        val id = UUID.randomUUID().toString()
-        val externalSignerLauncher = ExternalSignerLauncher("", signerPackageName = "")
-
-        val nostrSignerLauncher = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
-            if (result.resultCode != Activity.RESULT_OK) {
-                Log.e("Pokey", "ExternalSigner result error: ${result.resultCode}")
-            } else {
-                result.data?.let { externalSignerLauncher.newResult(it) }
-            }
-        }
-        externalSignerLauncher.registerLauncher(
-            launcher = {
-                try {
-                    nostrSignerLauncher.launch(it) // This can remain if you still need to launch it
-                } catch (e: Exception) {
-                    if (e is CancellationException) throw e
-                    Log.e("Pokey", "Error opening Signer app", e)
-                }
-            },
-            contentResolver = { Pokey.getInstance().contentResolverFn() },
-        )
-        externalSignerLauncher.openSignerApp(
-            "",
-            SignerType.GET_PUBLIC_KEY,
-            "",
-            id,
-        ) { result ->
-            val split = result.split("-")
-            val pubkey = split.first()
-            if (split.first().isNotEmpty()) EncryptedStorage.updatePubKey(pubkey)
         }
     }
 }
