+++ app/src/main/res/layout/fragment_home.xml
@@ -7,45 +7,64 @@
     tools:context=".ui.home.HomeFragment">
 
 
-    <EditText
-        android:id="@+id/npub_input"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:autofillHints="npub"
-        android:ems="10"
-        android:hint="@string/input_npub"
-        android:inputType="text"
-        android:minHeight="48dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.497"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
     <Button
         android:id="@+id/service_start"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="76dp"
+        android:layout_marginTop="68dp"
         android:text="@string/start"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@id/npub_input"
+        app:layout_constraintTop_toTopOf="@id/mainLayout"
         tools:layout_constraintTop_toBottomOf="@id/npub_input" />
 
     <ImageView
         android:id="@+id/imageView2"
         android:layout_width="109dp"
         android:layout_height="90dp"
-        android:contentDescription="Logo"
+        android:layout_marginTop="100dp"
+        android:contentDescription="@string/logo"
         android:src="@mipmap/ic_launcher"
-        app:layout_constraintBottom_toTopOf="@+id/npub_input"
+        app:layout_constraintBottom_toTopOf="@id/service_start"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.5"
         tools:ignore="MissingConstraints" />
+
+    <LinearLayout
+        android:id="@+id/mainLayout"
+        android:layout_width="286dp"
+        android:layout_height="50dp"
+        android:layout_marginTop="32dp"
+        android:orientation="horizontal"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/imageView2">
+
+        <EditText
+            android:id="@+id/npub_input"
+            android:layout_width="220dp"
+            android:layout_height="match_parent"
+            android:autofillHints="npub"
+            android:ems="10"
+            android:hint="@string/input_npub"
+            android:inputType="text"
+            android:minHeight="48dp" />
+
+        <Button
+            android:id="@+id/amber"
+            android:layout_width="57dp"
+            android:layout_height="48dp"
+            android:drawableTop="@drawable/ic_amber"
+            android:drawablePadding="22dp"
+            android:layout_marginStart="9dp"
+            android:gravity="center"
+            tools:ignore="MissingConstraints" />
+
+    </LinearLayout>
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
