+++ app/src/main/res/values/strings.xml
@@ -24,4 +24,5 @@
     <string name="relays">Your relays</string>
     <string name="start">Start</string>
     <string name="stop">Stop</string>
+    <string name="logo">Logo</string>
 </resources>
